requirements.txt
Django==1.9.5
requests==2.9.1
telepot==6.7
python-dotenv==1.0.1
gunicorn
supervisor
django-environ==0.4.0